# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.BudgetCreateInput:
    properties:
      currency:
        description: ISO 4217 (e.g., "BRL")
        enum:
        - BRL
        - USD
        - EUR
        example: BRL
        type: string
      items:
        items:
          $ref: '#/definitions/dtos.BudgetItemInput'
        type: array
      reference_month:
        description: YYYY-MM format
        example: 2025-01
        type: string
      total_amount:
        description: String decimal (e.g., "5000.00")
        example: "5000.00"
        type: string
    type: object
  dtos.BudgetItemInput:
    properties:
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      percentage_goal:
        description: String decimal (e.g., "25.50")
        example: "25.50"
        type: string
    type: object
  dtos.BudgetItemOutput:
    properties:
      budget_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      category_id:
        example: 880e8400-e29b-41d4-a716-446655440003
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 770e8400-e29b-41d4-a716-446655440002
        type: string
      percentage_goal:
        example: "30.00"
        type: string
      percentage_spent:
        example: "46.67"
        type: string
      planned_amount:
        example: "1500.00"
        type: string
      remaining_amount:
        example: "800.00"
        type: string
      spent_amount:
        example: "700.00"
        type: string
      updated_at:
        example: "2025-01-20T08:00:00Z"
        type: string
    type: object
  dtos.BudgetOutput:
    properties:
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      currency:
        enum:
        - BRL
        - USD
        - EUR
        example: BRL
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      items:
        items:
          $ref: '#/definitions/dtos.BudgetItemOutput'
        type: array
      percentage_used:
        example: "47.00"
        type: string
      reference_month:
        description: YYYY-MM
        example: 2025-01
        type: string
      spent_amount:
        example: "2350.00"
        type: string
      total_amount:
        example: "5000.00"
        type: string
      updated_at:
        example: "2025-01-20T08:00:00Z"
        type: string
      user_id:
        example: 660e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  dtos.BudgetPaginatedOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.BudgetOutput'
        type: array
      pagination:
        $ref: '#/definitions/dtos.BudgetPaginationMeta'
    type: object
  dtos.BudgetPaginationMeta:
    properties:
      has_next:
        example: false
        type: boolean
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJmIjp7InJlZmVyZW5jZV9tb250aCI6IjIwMjUtMDEifX0
        type: string
    type: object
  dtos.BudgetUpdateInput:
    properties:
      items:
        items:
          $ref: '#/definitions/dtos.BudgetItemInput'
        type: array
      total_amount:
        description: String decimal
        example: "6000.00"
        type: string
    type: object
...
  httperrors.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        additionalProperties: {}
        type: object
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Jailton Junior
    url: https://github.com/jailtonjunior94
  description: |-
    API para gestão financeira pessoal com suporte a cartões de crédito, transações,
    orçamentos mensais, categorias e faturas com parcelamento.

    ## Valores Monetários
    Todos os valores monetários são representados como **strings decimais** (ex: `"1234.56"`)
    para preservar precisão e evitar erros de ponto flutuante.

    ## Paginação
    Endpoints de listagem usam **cursor-based pagination**. Parâmetros:
    - `limit` (int, default: 20, max: 100)
    - `cursor` (string, opcional): valor de `pagination.next_cursor` da resposta anterior

    ## Autenticação
    Endpoints protegidos requerem o header `Authorization: Bearer {token}`.
    Obtenha o token em `POST /api/v1/token`.
  license:
    name: MIT
  title: Financial API
  version: 1.0.0
paths:
  /api/v1/budgets:
    get:
      description: Retorna a lista paginada de orçamentos do usuário autenticado (cursor-based
        pagination).
      parameters:
      - default: 20
        description: 'Itens por página (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Cursor de paginação
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de orçamentos
          schema:
            $ref: '#/definitions/dtos.BudgetPaginatedOutput'
        "400":
          description: Parâmetro inválido
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar orçamentos mensais
      tags:
      - budgets
    post:
      consumes:
      - application/json
      description: |-
        Cria um novo orçamento para o mês de referência do usuário autenticado.
        Apenas um orçamento por mês (`reference_month`) é permitido por usuário.

        **Campos:**
        - `reference_month`: `YYYY-MM` (ex: `2025-01`)
        - `total_amount`: valor total planejado (ex: `"5000.00"`)
        - `currency`: `BRL` | `USD` | `EUR` (opcional, default: `BRL`)
        - `items`: ao menos um item com `category_id` e `percentage_goal` (ex: `"25.50"`)
      parameters:
      - description: Dados do orçamento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.BudgetCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Orçamento criado
          schema:
            $ref: '#/definitions/dtos.BudgetOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "409":
          description: Orçamento já existe para este mês
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Criar orçamento mensal
      tags:
      - budgets
  /api/v1/budgets/{id}:
    delete:
      description: Remove um orçamento mensal e todos os seus itens. Esta operação
        é irreversível.
      parameters:
      - description: ID do orçamento
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Orçamento removido com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Orçamento não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remover orçamento mensal
      tags:
      - budgets
    get:
      description: Retorna os detalhes de um orçamento mensal específico, incluindo
        itens com valores planejados, gastos e percentuais.
      parameters:
      - description: ID do orçamento
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados do orçamento
          schema:
            $ref: '#/definitions/dtos.BudgetOutput'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Orçamento não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Buscar orçamento por ID
      tags:
      - budgets
    put:
      consumes:
      - application/json
      description: |-
        Atualiza o valor total e os itens de um orçamento existente.
        Os itens existentes são substituídos pelos novos itens enviados.
      parameters:
      - description: ID do orçamento
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.BudgetUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Orçamento atualizado
          schema:
            $ref: '#/definitions/dtos.BudgetOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Orçamento não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Atualizar orçamento mensal
      tags:
      - budgets
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: `Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"